# Childguard/docs/OpenAPI.yaml
openapi: 3.0.3
info:
  title: ChildGuard API
  description: API for managing organizations within the ChildGuard platform.
  version: "1.0.0"
servers:
  - url: https://api.childguard.com/v1
...
paths:
...
  /user/login:
    get:
      tags:
        - Users
      summary: Logs user into the system
      description: ""
      operationId: loginUser
      parameters:
        - name: username
          in: query
          description: The user name for login
          required: false
          schema:
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid username/password supplied
...
components:
  schemas:
...
    User:
      type: object
      properties:
        userId:
          type: string
          description: Unique identifier for the user.
          example: "user123"
        username:
          type: string
          description: Username of the user.
          example: "johndoe"
        email:
          type: string
          format: email
          description: Email address of the user.
          example: "johndoe@example.com"
        role:
          type: string
          description: Role of the user within the platform.
          example: "volunteer"
...